Moved generator
