+++ source/generators/response.generator.ts
@@ -1,7 +1,7 @@
 import { MessageGenerator } from "./abstract.message.generator";
-import { IResponseReject, IResponseSuccess } from "../../../validators/interfaces/response.interfaces";
+import { IResponseReject, IResponseSuccess } from "../validators/interfaces/response.interfaces";
 import { IIterableDataField } from "./generator.interfaces";
-import { ValidatorService } from "../../../validators/validator";
+import { ValidatorService } from "../validators/validator";
 
 export class ResponseGenerator extends MessageGenerator<IResponseSuccess | IResponseReject> {
     private validator: ValidatorService = new ValidatorService()
